# backend/openapi.yaml
openapi: 3.1.0
...
info:
  title: LoopBuy Backend API
  version: 1.0.0
  description: >-
    REST API for LoopBuy accounts, marketplace listings, social features,
    scam screening, recommendations, and the RAG shopping assistant.
servers:
  - url: http://localhost:8090
    description: Docker Compose local API
security: []
...
paths:
...
  /api/v1/categories:
    get:
      tags: [Categories]
      operationId: listCategories
      summary: List active categories
      responses:
        '200':
          description: Active categories in display order.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CategoryList'
        default:
          $ref: '#/components/responses/Problem'
    post:
      tags: [Categories]
      operationId: createCategory
      summary: Create a category
      description: Administrator role is required.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CategoryCreateRequest'
      responses:
        '201':
          description: Category created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Category'
        default:
          $ref: '#/components/responses/Problem'
  /api/v1/categories/{identifier}:
    parameters:
      - $ref: '#/components/parameters/CategoryIdentifier'
    get:
      tags: [Categories]
      operationId: getCategory
      summary: Get an active category by ID or slug
      responses:
        '200':
          description: Category details.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Category'
        default:
          $ref: '#/components/responses/Problem'
    patch:
      tags: [Categories]
      operationId: updateCategory
      summary: Update a category
      description: Administrator role is required and `identifier` must be numeric.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CategoryPatchRequest'
      responses:
        '200':
          description: Updated category.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Category'
        default:
          $ref: '#/components/responses/Problem'
    delete:
      tags: [Categories]
      operationId: deleteCategory
      summary: Deactivate a category
      description: Administrator role is required and `identifier` must be numeric.
      security:
        - bearerAuth: []
      responses:
        '204':
          description: Category deactivated.
        default:
          $ref: '#/components/responses/Problem'
...
  /api/v1/users/{id}:
    parameters:
      - $ref: '#/components/parameters/Id'
    get:
      tags: [Users]
      operationId: getPublicUser
      summary: Get a public user profile
      responses:
        '200':
          description: Public profile without email, role, status, or phone.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        default:
          $ref: '#/components/responses/Problem'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: HS256 access token issued by register, login, or refresh.
  parameters:
    Id:
      name: id
      in: path
      required: true
      schema:
        type: integer
        format: int64
        minimum: 1
    CategoryIdentifier:
      name: identifier
      in: path
      required: true
      description: Numeric category ID or slug for GET; numeric ID for PATCH/DELETE.
      schema:
        type: string
        pattern: '^(?:[1-9][0-9]*|[a-z0-9]+(?:-[a-z0-9]+)*)$'
...
  responses:
    Problem:
      description: Error response. Typical statuses are 400, 401, 403, 404, 409, 422, 428, 500, or 503.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
  schemas:
...
    Problem:
      type: object
      additionalProperties: false
      required: [type, title, status]
      properties:
        type:
          type: string
          format: uri
        title:
          type: string
        status:
          type: integer
          minimum: 400
          maximum: 599
        detail:
          type: string
        instance:
          type: string
        errors:
          type: object
          additionalProperties:
            type: string
...
    CategoryCreateRequest:
      type: object
      additionalProperties: false
      required: [name, slug]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 100
        slug:
          type: string
          maxLength: 100
          pattern: '^[a-z0-9]+(?:-[a-z0-9]+)*$'
        is_active:
          type: boolean
          description: Accepted but ignored on creation; new categories are active.
    CategoryPatchRequest:
      type: object
      additionalProperties: false
      properties:
        name:
          type: string
          maxLength: 100
        slug:
          type: string
          maxLength: 100
          pattern: '^[a-z0-9]+(?:-[a-z0-9]+)*$'
        is_active:
          type: boolean
...
    Profile:
      type: object
      additionalProperties: false
      required: [user_id, full_name, phone, location, bio, profile_image, updated_at]
      properties:
        user_id:
          type: integer
          format: int64
        full_name:
          type: string
        phone:
          type: string
          description: Empty in public-user responses.
        location:
          type: string
        bio:
          type: string
        profile_image:
          type: string
        updated_at:
          type: string
          format: date-time
    User:
      type: object
      additionalProperties: false
      required: [user_id, username, email_verified, created_at]
      properties:
        user_id:
          type: integer
          format: int64
        username:
          type: string
        email:
          type: string
          format: email
          description: Present only in private-user responses.
        email_verified:
          type: boolean
          description: False in public-user responses; authoritative in private/session responses.
        role:
          type: string
          enum: [user, moderator, admin]
          description: Omitted in public-user responses.
        status:
          type: string
          enum: [active, suspended, deleted]
          description: Omitted in public-user responses.
        created_at:
          type: string
          format: date-time
        profile:
          $ref: '#/components/schemas/Profile'
    Category:
      type: object
      additionalProperties: false
      required: [category_id, name, slug, is_active, created_at]
      properties:
        category_id:
          type: integer
          format: int64
        name:
          type: string
        slug:
          type: string
        is_active:
          type: boolean
        created_at:
          type: string
          format: date-time
...
    CategoryList:
      type: object
      additionalProperties: false
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Category'
...